backend/requirements.txt
blinker==1.8.2
click==8.1.7
dotenv==0.9.9
flask==3.0.3
Flask-Cors==5.0.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==2.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing-extensions==4.12.2
werkzeug==3.0.6
zipp==3.20.2
